Added orphans cache size in the configuration of the local nodes

+++ io-mokamint-node-messages/src/main/java/io/mokamint/node/messages/WhisperedMemories.java
@@ -33,7 +33,7 @@ public final class WhisperedMemories {
 	 * @return the memory
 	 * @throws IllegalArgumentException if {@code size} is negative
 	 */
-	public static WhisperingMemory of(long size) {
+	public static WhisperingMemory of(int size) {
 		return new WhisperedMemoryImpl(size);
 	}
 }
